multivm.MemoryChromeOSPerf test
Like the arc.MemoryChromeOSPerf test, this test allocates memory until
ChromeOS is at a certain memory pressure level. Things that are new:
* Uses metrics.LogMemoryStats for detailed metrics, including the size
  of VMs.
* Applies memory pressure in three steps, idle (no pressure), moderate,
  and critical.

This allows us to measure the memory cost of features a different levels
of memory pressure. Since swap and reclaim (etc.) can change how much
memory a process uses, it is useful to measure costs at different memory
pressure levels to see how it responds.

arc.MemoryChromeOSPerf will be removed once we have a history of data in
crosbolt for this new test.

BUG=b:220688806
TEST=tast run dut multivm.MemoryChromeOSPerf.with_bg_arc

Change-Id: Ib0c66e868e353b4dfe84b793f5cfa2849fe04e5a