cryptohome: Trim spaces on parsing auth_session_id
Strip leading and trailing whitespaces when parsing the
"auth_session_id:..." line in the cryptohome CLI tool output.

This change is necessary as we're going to switch the CLI to print a
whole protobuf instead of a single line, and the protobuf's standard
pretty printer emits a space after the colon.

BUG=b:208357938
TEST=cherry-pick crrev.com/c/3462649 and run "tast run $IP cryptohome.PersistentAuthSession"

Change-Id: Ib0f1d4e39c28b6f7d6c4d4a1f504bbb14fb721e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3463675
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>