wifi: Passpoint: provide a simulated environment.
The CL provides a simulated Wi-Fi environment implemented as a fixture.
The fixture is based on the mac80211_hwsim kernel module to create the
simulated environment and obtain Wi-Fi test interfaces. Shill is then
configured to ignore the hardware interface (if any) and use one of the
test interfaces. A test that depends on the fixture receives a struct
that describes the test environment.

BUG=b:162258594
TEST=tast run ${DUT} wifi.PasspointSelection.* wifi.PasspointRoaming.*
(coming in subsequent CLs)
tast run ${DUT} wifi.WPAValidity wifi.UdevRename wifi.Regulatory wifi.Driver \
    wifi.IWScan

Change-Id: Id652dec6482fa7fcdd2a680c47fa20de1ed62981
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3610629
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Damien Dejean <damiendejean@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>