callbox: callbox manager client, fixture, and example test
These changes allow for tast tests to be written using a new fixture,
callboxManagedFixture, that utilizes the new Callbox Manager service
on the chromeos1-proxy server to manage connected callboxes. The
cellular_Callbox_AssertCellularData autotest test has been recreated
for tast using this new fixture as an example.

The test described in TEST currently fails, but this is due to an
actual failure during BeginSimulation due to an issue with the test
bed. The test itself does successfully interface with the callbox
manager and should pass once the test bed is fixed.

BUG=b:219962821
TEST=tast run \
  -var=callbox=$CALLBOX \
  -var=callboxManager=$CALLBOX_MANAGER \
  $DUT \
  cellular.AssertCellularData

Change-Id: I01b3067b737ab29f92646c641a31845ea29963da