typec: Use wakeup sysfs class to check cros-ec wake count
Let's use the wakeup sysfs path to verify the wakeup count increases vs.
parsing the debugfs file. This is more future proof as sysfs is an ABI
and it also avoids having to look for a specific device name in the
wakeup debugfs file. Furthermore, it fixes this test on ARM devices that
don't use the ACPI name because cros-ec probes on a different bus (i2c,
spi, remoteproc). We have to navigate sysfs and we take the path of
finding the parent device of the cros_ec device as that's the device
used for the wakeup accounting.

BUG=b:220040007
FIXES=b:220040007
TEST=tast run --var=servo=localhost $trogdor_dut typec.HpdWake

Change-Id: Ie562c4930f770afe2a23fdc25a9f4d56c6f9585d