tast-tests: Stabilize policy.LacrosMainProfileLogin
The test is failing because the sync server responds with "too many
requests". Add a pool of accounts to be used for login so that we have
less requests per account.

Remove accountmanager.LacrosMainProfileLogin test because it tests the
same behavior and due to complex test setup (requiring a separate
account pool) needed for the fix.

BUG=b:215674255
TEST=tast run <ip> policy.LacrosMainProfileLogin

Cq-Depend: chrome-internal:4554962
Change-Id: I35cb0ebf61891b77d7f008d689d895b33bc162ef