policy: Fix flakiness of FileSystemWriteBlockedForUrls test
It looks like sometimes the test ended while Chrome was still writing
the file to disk, which caused the file to not get cleaned up properly.
I also extended the test to actually write some content to the file and
verify that the file contains the content.

BUG=chromium:1297527
TEST=tast -verbose run <IP> policy.FileSystemWriteBlockedForUrls*

Change-Id: I9a0136e3d2163fb6207888ed370585feb5cd65ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3470102
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Commit-Queue: Christian Flach <cmfcmf@chromium.org>
Tested-by: Christian Flach <cmfcmf@chromium.org>