tast-tests: Add lacros variant to ui.TabletOperations
The following test needs Lacros enabled, and test Javascript exception
in browser.
- ui.TabletOperations

Note that there seems to be a change of behavior on tapping on the
browser icon on the app-list.
It opens a new tab for ash-chrome, but a new window for lacros-chrome,
which may affect some perf metrics if it has to be compared between
ash/lacros.

See go/lacros-tast-porting for the porting guidelines.

BUG=b:216242056
TEST=tast run ui.TabletOperations*
TEST=tast run '("group:mainline" && !"dep:lacros" && "name:ui.*")'  # ash-chrome tests

Change-Id: I71c2c644229ea1d6a035ed9049d9ed49dbb2367d