tast: Add error detection in arc.OptinManaged
Currently the test succeeds if the Play Store launches irrespective of
the presence of an error on the window. With this change, we're
introducing a helper method that can be used to ensure that there is no
error on display in the Play Store window.

BUG=b:219977901
TEST=tast run DUT arc.OptinManaged.vm
TEST=tast run DUT arc.Optin.vm

Change-Id: If8467f38aa2b9d089a2dfc73f5ee9067b2034d85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3473879
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Auto-Submit: Muhammad Hasan Khan <mhasank@chromium.org>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Commit-Queue: Judy Wu <jinrongwu@google.com>