webrtc: Set a fixture of enabling vp9 k-SVC hw decoding
VP9 k-SVC hw decoding has been launched on x86 devices. We are going to
release the feature on trogdor. The feature is disabled by default on
the device and will be tested via finch experiment. Therefore, to run
the test decoding vp9 k-SVC stream on trogdor too, this CL specifies a
fixture of enabling vp9 k-SVC hw decoding.

BUG=b:190629171
TEST=webrtc.RTCPeerConnection.vp9_dec* on trogdor

Change-Id: I9fff751b717a06e67548d004571431305b707cbf