tast-tests: Move rollback_service to autoupdate
Relocate rollback_service and its corresponding proto to autoupdate.
Modify all dependencies and references to the service and proto and
test the affected tests build and run as expected.

Autoupdate contains the services related to rollback and have access
to the package lsbrelease.go, which we need to use in this service in
an ongoing update.

BUG=chromium:1264389, b:217546714
TEST=tast -verbose run DUT policy.EnterpriseRollbackInPlace passes
tast run -var=updateutil.tlwAddress="localhost:7151" localhost:2200
autoupdate.EnterpriseRollbackPreviousVersion runs as expected

Change-Id: I581a4be028405f3c62668545e2243c90a770ae41