tast-tests: Mimic session end for cryptohome unmounts
Make cryptohome unmounts as similar to the session end as possible.
We use process killer to ensure that no process is writing to a
file before the unmount and no process is holding a mount open
post unmount.

Note: Chapsd on brya seems to be frequently holding up the unmount
process (and this doesn't seem like a problem on other devices
at the moment). Until this is fixed, ensure that chapsd restarts
and is running correctly post unmount.

BUG=b:214879437
TEST=tast run hwsec.Chaps*, hwsec.AccountDiskUsage passes

Cq-Depend: chromium:3446463
Disallow-Recycled-Builds: test-failures
Change-Id: Ieb80284829b878ecc5cd719b8770bf0a6645aa43