tast-tests: Simplify ui.ChromePIPEnergyAndPower.*
After resizing the PIP window, the tests move the mouse to ensure that
the PIP window will show no controls or resize shadows. That is
unnecessary because there is no way such controls or resize shadows will
persist as the test opens a second browser tab and presses the Tab key.

BUG=chromium:1093933
TEST=tast run <DUT> ui.ChromePIPEnergyAndPower.*

Change-Id: I2c0c53e443ed58e1a116fb963b0b9d8f553d775a