wifi.SetTXPower: Correctly expect failure in unsupported modes
See ${BUG} for a deep dive into the context of this fix.

TL;DR: This test currently has a special exception that expects all
invocations of `set_wifi_transmit_power` with `--domain=none` to
complete successfully, even when the tablet or non-tablet mode is
unsupported. This expectation was based on a faulty assumption, and is
causing failures in our testing labs. This change removes this special
exception.

BUG=b:213383779
TEST=tast run DUT wifi.SetTXPower
Passes on a variety of Spherion DUTs, including those that were
previously failing.

Change-Id: I603f04b7b2c7984ebc80a0da43ce4e796a24fd78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3475596
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Auto-Submit: Kevin Lund <kglund@google.com>
Reviewed-by: Jun Sheng <shengjun@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>