tast-tests: Make PIP big in ChromePIPRoundedCornersUnderlay
ui.ChromePIPRoundedCornersUnderlay* will resize the PIP window to
maximum size, to ensure that the video does not have to be supersampled
as this may affect the overlay strategy. Also, on devices that default
to tablet mode, the PIP button will be tapped instead of clicked.

BUG=chromium:1290221
TEST=tast run <DUT> ui.ChromePIPRoundedCornersUnderlay*

Change-Id: I406994a5b38de534e1072b92e7970698b2ab31b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3475595
Tested-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Peter McNeeley <petermcneeley@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Eliot Courtney <edcourtney@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>