tast-tests: Simplify arc.PIPEnergyAndPower.big*
After resizing the PIP window, the tests move the mouse to ensure that
the PIP window will show no controls or resize shadows. That is
unnecessary because there is no way such controls or resize shadows will
persist as the test opens a browser tab and presses the Tab key.

BUG=chromium:1093933
TEST=tast run <DUT> arc.PIPEnergyAndPower.big*

Change-Id: I871623105333f3ec711d222e5eb83ea628e05a7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3474349
Tested-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Achuith Bhandarkar <achuith@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>