mta: Enhancement of launcher.AppDragAndDrop
Previously the launcher pages are all identical on the UI tree, the old way to identify where a launcher item is located was based on it.
This CL removes the old way, adopts a simpler way to detect a launcher item due to the UI was updated.

BUG=b:223292711
TEST=tast run <DUT> launcher.AppDragAndDrop

Change-Id: I1bf0a21d04c96b6333d6ed2766b7c5fe229fbf3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3531892
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: David Welling <davidwelling@google.com>
Tested-by: Ting Chen <ting.chen@cienet.com>
Commit-Queue: Ting Chen <ting.chen@cienet.com>