tast-tests: Skip VerifyRollback steps if expected failure
There are a few steps in VerifyRollback that are expected
to fail if the milestone of the image is <100. Skip these
steps in those cases and include a corresponding failure
message to inform that the failure is expected.

Extending the proto to be able to provide details about the
failure should also be useful in the future when extending
rollback_service to make a distinction between different
possible failures raised during the verification process.

Update the two tests affected by this modification to
include FailureDetails in the error message if provided.

BUG=chromium:1264389, b:217546714
TEST=tast -verbose run DUT policy.EnterpriseRollbackInPlace passes if
milestone>=100
tast -verbose run DUT policy.EnterpriseRollbackInPlace fails
if milestone<100 and error message shows expected failure details
tast run -var=updateutil.tlwAddress="localhost:7151" localhost:2200
autoupdate.EnterpriseRollbackPreviousVersion shows failure details

Change-Id: I1f3078d237930733abf110c79864b86315e7afc0