tast-tests: Skip VerifyRollback steps if expected failure
There are a few steps in VerifyRollback that are expected
to fail if the milestone of the image is <100. Skip these
steps in those cases, respond with success and include a
message to inform that the steps have been skipped.

Extending the proto to be able to provide details about the
verification is useful for this specific case but should also
be useful in the future when extending rollback_service to make
a distinction between different possible failures raised during
the verification process.

Update the two tests affected by this modification to
include VerificationDetails in the error message and log these
if the verification is successful but the details are provided.

BUG=chromium:1264389, b:217546714
TEST=tast -verbose run DUT policy.EnterpriseRollbackInPlace passes
tast run -var=updateutil.tlwAddress="localhost:7151" localhost:2200
autoupdate.EnterpriseRollbackPreviousVersion passes and log shows
details

Change-Id: I1f3078d237930733abf110c79864b86315e7afc0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3478194
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>