Crostini: Test unsharing folder when file opened in Gedit.
This test uses uidetection (acuiti) to automate testing of unsharing
folders with crostini. This test automates this manual test [1]. A text
file is created, and opened with Gedit before the folder is un-shared. As
of now this does not does a screendiff, but once I work out how to reliably
have the error banner show up on DUTs I will add screendiff.

[1] https://testtracker.googleplex.com/testplans/testcase/detail/2208021?id=18920&revision=199

BUG=b:220779813
TEST=Ran on hatch (Akemi) DUT and remotely on lab Atlas boards.

Change-Id: I2926c7fdfd946780cb6e3b60e37742a369b6a640
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3478955
Reviewed-by: Judy Wu <jinrongwu@google.com>
Reviewed-by: David Munro <davidmunro@google.com>
Tested-by: Zubin Pratap <zubinpratap@google.com>
Commit-Queue: Zubin Pratap <zubinpratap@google.com>