tast-tests: Update drive oauth flow and migrate to fixtures
In crrev.com/c/4567092 the refresh token was added to the authentication
credentials. According to the Drive API documentation [1] the refresh
token never expires, so this takes advantage of this to retrieve an
access token and get rid of the brittle auth code flow.

The following additional items have been completed as well:
 - Migrate the drivefs precondition to a fixture
 - Remove dats@ from Contacts as they are not with the Files team
 - Update the LaCrOS status for the Drive related tests

[1] https://developers.google.com/identity/protocols/oauth2#expiration

BUG=chromium:1299712
TEST=tast run <DUT-IP> filemanager.Drivefs*

Change-Id: I1c7e2b09ee4eb1e12adadc1448e13e6c3af8d127