tast-tests: Move board check to beginning of test
The test should skip and pass for unreleased boards. Move this check
at the beginning of the test before any other step is carried out.

When the test is skipped, the corresponding rollback cleanup is
unnecessary, so the test should finish before this cleanup is deferred
to avoid errors.

BUG=chromium:1264389, b:217546714
TEST=tast run -var=updateutil.tlwAddress="localhost:7151" localhost:2200
autoupdate.EnterpriseRollbackPreviousVersion passes for unreleased board
and message appears in the log.

Change-Id: I6fa0989739f9762856074037435b051b067b6933
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3479862
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>