typec: Disable basic test on habokay
Habokay seems to run into an error where we can't perform an "ls" command
after a reconnection to the DUT is established. This is despite the fact
that on the DUT itself, the file is clearly present.

While this error should be resolved, it doesn't point to any issue with
the Type C subsystem itself, so let us blocklist habokay (other keeby
variants continue to work).

BUG=b:217955802
TEST=Use a crosfleet habokay and verify that the test is skipped.

Change-Id: I3198d21a6e3a6b4bad763157f4fdededef36d3c3