Increase optin timeout, remove redundant recovery
In Change-Id I50def1b984f377259cf636427111aab7ead341b0 we introduced a possibility to retry play optin in case of an error. This retry should not be a part of Perform func, as it is expected to not have retries. As a part of refactoring, we introduced and extra recovery step, which is redundant.

Remove retry for opt-in. Do not perform recovery steps, if the action failed maxAttempts times. Extract recovery steps in a separate function.

BUG=b:221112204
TEST=tast run dut arc.OptinNetworkError.vm

Change-Id: I0b68c2c4c2e7e0bc60decb8832feb63ee9890a66
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3486816
Tested-by: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: Muhammad Hasan Khan <mhasank@chromium.org>
Reviewed-by: Judy Wu <jinrongwu@google.com>
Commit-Queue: Alexey Shpakov <ashpakov@google.com>