Increase optin timeout, remove redundant recovery
In Change-Id I50def1b984f377259cf636427111aab7ead341b0 we introduced a possibility to retry play optin in case of an error. However, the timeout was left the same. As a part of refactoring, we introduced and extra recovery step, which is redundant.

Double the timeout for the optin. Do not perform recovery steps, if the action failed maxAttempts times.

BUG=b:221112204
TEST=tast run dut arc.OptinNetworkError.vm

Change-Id: I0b68c2c4c2e7e0bc60decb8832feb63ee9890a66