u2fd: Migrate all usages of securitykeys.info to webauthn.io
Currently we use securitykeys.info for simple tests that do not need to
configure authenticator type and user verification preference. But now
it's not useful anymore because its supported public key types cannot be
configured and it only supports ECC, but we're starting to use RSA keys
in TPM 1.2 devices.

Replace the utility function that performs WebAuthn in securitykeys.info
to a function that performs WebAuthn in webauthn.io, and change the
existing usages.

BUG=b:210418148
TEST=tast run $DUT u2fd.WebauthnUsing* #on tpm1 and tpm2 devices

Change-Id: Ib2fc60ccf29af3822c0acdcc5e88530573ef282b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3487195
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>