tast-tests: MeetCUJ click away permission bubble for 2 min
Waiting for a fixed set up permissions makes it unnecessarily
complicated because of the behavior change of permission bubble
for "notification" with or without finch test config. The CL
makes it simple that MeetCUJ just clicks away any permission
bubble for 2 minutes. The side effect of doing that is the extra
1 minute wait after all permissions are granted.

BUG=b:219281373
TEST=tast run $DUT ui.MeetCUJ.16p with finch test config

Change-Id: I3c0c4d15963e203ded02bcbe275a978f374cbec4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3488556
Reviewed-by: Yichen Zhou <yichenz@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>