tast-tests: Increase enterprise.ARCProvisioning timeout for Play Store
The timeout currently set to 1 minute is not sufficient. UMA stats show
that p99 takes 1 minute 30 seconds where as the timeout set in code is 3
minutes.

The arc.OptinManaged test has the timeout set to 4 minutes and passes
consistently on the same build that this test failed.

Also removed unused constants in the test.

BUG=b:186274474
TEST=tast run DUT enterprise.ARCProvisioning.vm

Change-Id: Idec43dd3c0faef49793ef7a49e92c3b4ead1bc5e