tast-test: Update policy.UserFeedbackAllowed test
The test is flaky because it is not always able to find and open
the Help option in the menu. It is not visible in all screen sizes,
so we make sure that it is visible before attempting to click on it.

The screenshots captured during faillog show the popup window
resulting from the key_combination test instead of showing the
browser. Make sure as well that the popup window is closed so
menu_access test starts with a clean state.

BUG=chromium:1297117
TEST=tast run DUT policy.UserFeedbackAllowed* passes

Change-Id: If6b22352ca4df55411910c3ebe87e9eac37a56ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3491005
Reviewed-by: Kamil Szarek <kamilszarek@google.com>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>