tast-tests: Allow flaky ARC resizing in WindowArrangementCUJ
I want to extend WindowArrangementCUJ tests with ARC window resizing
workflows, but this can cause test flake because ARC window bounds
sometimes stop updating during a resize (see Comment 3 at [1]).
Fortunately, the bounds still update properly when the drag ends.

WITHOUT THIS CL: If window bounds stop updating during a resize in a
WindowArrangementCUJ test, then the test will fail because it waits for
the expected bounds before ending the drag.

WITH THIS CL: If a WindowArrangementCUJ test fails to wait for the
expected window bounds before ending a drag, it will log a warning, but
if the expected bounds are achieved after ending the drag, then the test
will proceed without failing.

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=1297297#c3

BUG=b:208290911, chromium:1297297
TEST=tast run <DUT> ui.WindowArrangementCUJ.*

Change-Id: I6b3a50cdb14f6ee70b068affc6661916b9b3a585