phonehub: add test for recent Chrome tabs
Add a test that verifies Phone Hub updates to show recent Chrome tabs
from the Android device, and that clicking the recent tab chips launches
Chrome with the correct URLs.

Chrome sync is enabled on Android using UIAutomator. If we clear all of
Chrome's data before launching it, there is a streamlined flow to enable
it. First there is a page to accept the ToS. Accepting the ToS then
conveniently leads to an opt-in page for Chrome Sync, so we only need to
click 2 buttons total on Android to set it up.

BUG=b:221495526
TEST=tast run <dut> phonehub.*

Change-Id: I36519ae5ea29a395c4bc798c5229b0fd885356d7