samba: Add filemanager.SMBPassword.remember_password test
SmbFs has the ability to take a username and password to authenticate
against a Samba share. It's possible to have the password be remember
after a reboot. Verify if the user has "Remember my password" checked
the files are immediately available and the user is not prompted for a
password.

This adds a fixture that expects the test to cleanup Chrome as if the
test is to restart Chrome is changes the fixture's Chrome instance so it
can't appropriately cleanup. Ensure this is explicit in the fixture
description and the test does adequate cleanup.

The smbpasswd file provides a password for the chronos user that
overrides the password in /etc/passwd. This is set as test0000 and the
Samba server ensures to read from this file to authenticate the user.
This file was generated using smbpasswd -c smb.conf -a chronos

BUG=chromium:1295640
TEST=tast run <DUT-IP> filemanager.SMBPassword*

Change-Id: Ia6d38789895e5eec9c4bf0d97f98b26122e29c32