lacros: add variance and max latency to PowerVideocall
Always good to report a variance when reporting a mean, so we
approximate the variance from the histogram by computing the midpoint of
each histogram bucket. There also seems to be some spurious maximum
latency values being reported by LaCrOS, so we add that statistic as
well, for tracking.

BUG=chromium:1301340
TEST=Manual testing on crosfleet devices.

Change-Id: I22cceb3db9dd889126dbd79f69f0b82334a587d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3495202
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Luke Nihlen <luken@google.com>
Tested-by: Luke Nihlen <luken@google.com>