camera: Lengthen the timeout for testPreviewPerformance
For devices with two cameras, sometimes it takes about 1 minutes to wait
for the CPU to idle and takes at least 25 * 2 (Before/After QRCode
detection enabled) * N (Amount of camera). Therefore, the current
timeout 3 minutes is a little too strict for the test to pass and cause
many test failures.

This CL lengthen the timeout for testPreviewPerformance to 5 minutes.

BUG=b:222229057
TEST=None

Change-Id: Ie4ad3f1eef095386e85168778b739e9c28bfde9d