tast-tests: Cleanup before starting attestation test
Fixup cleanup in AttestationNoExternalServer. CQ runs with the test
show that due to a timing condition on VMs running on GCE,
session_manager and chrome still have files open when the test
attempts to cleanup existing state and hence get killed.

Instead, move cleanup to the beginning of the test.

BUG=b:214879437
TEST=hwsec.AttestationNoExternalServer passes on VM

Change-Id: I3b914a58ce4524ee6bdae71501e6c7c9d840705b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3499546
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>