(On hold) refactor: browserfixt.SetUpWithNewChrome for tests with no lacros fixture
DO NOT SUBMIT (on hold as edcourney@ working on another approach)

Currently lacros fixture is required to setup lacros for tests.
However, some ash-chrome tests that need lacros variant don't have any
fixture at all but simply create a new chrome instance in test body.

To support this scenario, some of the setup logic in lacrosfixt needs
to be refactored then exposed for other package to use.

This CL will do:
- refactor lacrosfixt.SetUp to expose common utils for other tests
or any package to use
- introduce a new browserfixt.SetUpWithNewChrome
- use it to add a lacros variant to holdingspace.Download for
a proof-of-concept

BUG=b:222242545
TEST=Use rootfs lacros binary by default
TEST=  tast run ${DUT} holdingspace.Download*
TEST=Use deployed binary with the var
TEST=  tast run -var=lacrosDeployedBinary=/usr/local/lacros-chrome ${DUT} holdingspace.Download*
TEST=Run CQ regression checks
TEST=  tast run ${DUT} ("group:mainline" && "dep:lacros" && !informational)
TEST=  tast run -var=lacrosDeployedBinary=/usr/local/lacros-chrome ${DUT} ("group:mainline" && "dep:lacros" && !informational)

Change-Id: Ib1fb3c140844323def9c21dc3c25ffd6581f29c8