refactor: browserfixt.SetUpWithNewChrome for tests with no lacros fixture
Currently lacros fixture is mandated to setup lacros for tests.
However, some ash-chrome tests that need lacros variant don't have any
fixture at all but simply create a new chrome instance in test body.

To support this scenario, some of the setup logic in lacrosfixt needs
to be refactored then exposed for other package to use.

This CL will do:
- refactor lacrosfixt.SetUp to expose common utils for other tests
or any package to use
- introduce a new browserfixt.SetUpWithNewChrome
- use it to add a lacros variant to holdingspace.Download for
a proof-of-concept

BUG=b:222242545
TEST=tast run holdingspace.Download.lacros_cancel

Change-Id: Ib1fb3c140844323def9c21dc3c25ffd6581f29c8