[TAST] Remove references to IconButton class.
This change is a follow up change to update references to the IconButton
class to instead remove references to the IconButton class. The existing
node finder is specific enough to find the button. This is a much better
approach than searching for the class name with a different mechanism
because we no longer rely on a hardcoded C++ class name.

BUG=b:218882177
TEST=CQ and ran on device.

Change-Id: I8f79bc960a2ac9e29113ba7a4e4d287b8e17568e