[TAST] Fix references to IconButton class.
This change is a follow up change to update references to the IconButton
class to instead reference FeaturePodIconButton. We have to update the
class name that we search for and cannot rely on HasClass since HasClass
internally will search for the class name provided, expecting the class
name provided to be found between word boundaries ("\b").

BUG=b:218882177
TEST=CQ and ran on device.

Change-Id: I8f79bc960a2ac9e29113ba7a4e4d287b8e17568e