playstore: Add ui.Enabled(true) for link your PayPal account, Complete account setup messages
BUG=b:217964717
TEST=tast run <helios> arcappcompat.Snapchat.vm_clamshell_mode

In the current version of Playstore.go, it does not check if ui.enabled
is true for "Want to link your PayPal account" and “Complete account setup”
text messages. Observed that the “Complete account setup” block is executed
even though “Complete account setup” is not in the foreground.  So added
ui.Enabled(true)check for "Want to link your PayPal account" and “Complete
account setup”  text message so that inside the block inside "Want to link
your PayPal account" and “Complete account setup” will be executed only
if they are enabled. This will help fix the flaky installation failures.

Change-Id: I9a94da965a0850819a7861181bcf3fe6bf2c21a2