Don't click at display edge when trying to close the launcher
launcher.BubbleSmoke test clicks at display origin to close the launcher
bubble widget. This attempt to click on display edge has lately started
crashing on some boards because the a target location cannot be assigned
to a window. This moves the target click close to top left edge, but not
exactly at origin, to ensure the target location is within a root
window bounds.

BUG=b:221688041
TEST=tast run dut launcher.BubbleSmoke* on kukui-arc-r

Change-Id: Id8467184303aa8ee613960c59b02ebb95ef7e896