Increase ui timeout
appsplatform.CheckSkuPurchase tests are frequently failing on the slower machines due to a ui timeout (failure screenshots show the elements, that were missing, or a waiting spinner). At the same time, we are using two different timeouts in the test, 10s and 30s.

Unify timeouts to be 30s throughout the test. This will increase previously failing 10s timeouts to 30s and should stabilise the test on slower machines.

BUG=chromium:1290626
TEST=tast run dut appsplatform.CheckSkuPurchase.vm

Change-Id: I640657618a072c775d27b106ea06f8174ca9d27a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3499253
Tested-by: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Jeevan Shikaram <jshikaram@chromium.org>
Commit-Queue: Alexey Shpakov <ashpakov@google.com>