firmware: ECSharedMemory() test case
This testcase uses the Servo V4+ ability to check if enough free space
is available in shared memory.
The test scenario is as follows, first the amount of free shared memory
in the EC is checked when the chromebook is in normal mode, then the EC
crash is forced after which the amount of available shared memory is
checked again. Finally it checks if the EC has made a jump to RW if not
then it will make one and the available free space in shared memory is
checked already in RW.

This test replaces firmware_ECSharedMem.py from FAFT.

BUG=b:194908259
TEST=tast run -var servo=${Servo_IP} ${DUT_IP} firmware.ECSharedMemory

Change-Id: I3216b12188df9dce16495859fa91f01d7e041364
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3497468
Reviewed-by: Matt Stark <msta@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Kshitij Shah <tij@google.com>
Reviewed-by: Jakub Skrzypnik <js@semihalf.com>
Reviewed-by: Przemysław Filipek <pf@semihalf.com>
Tested-by: Przemysław Filipek <pf@semihalf.com>
Commit-Queue: Przemysław Filipek <pf@semihalf.com>