tast-tests: Autogenerate nobatterymetrics variants
As a step toward collecting power metrics in all (or almost all) CUJ
tests, this CL provides a way to generate parameters of tests that only
run on devices supporting forced battery discharge, and corresponding
nobatterymetrics variants which only run on devices that do not support
forced battery discharge. A related idea was proposed in [1] and was
unfavored because "generating parameters at runtime would make it
difficult to tell which tests are defined by reading code" (Comment 2),
but for a different use case. I want a simple pattern of parameters
consistent across a large number of tests, whereas the stated purpose of
[1] is for tests individually having complicated patterns of parameters.

Although the primary purpose of this CL is to implement test parameter
generation functionality for use in CUJ tests, this CL modifies some ARC
tests to use the functionality because those are the tests that already
follow the pattern of nobatterymetrics variants. It is just refactoring.

See [2-6] for examples of what I will have to do with each CUJ test if I
apply the technique that was favored in [1]. Compare that to this CL.

This CL uses the functionality of [7]. It should be landed after [7].

[1] https://crbug.com/1065846
[2] https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3508414
[3] https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3508778
[4] https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3508774
[5] https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3508777
[6] https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3508773
[7] https://chromium-review.googlesource.com/c/chromiumos/platform/tast/+/3503726

BUG=b:218879803
TEST=../tast/fast_build.sh -T

Change-Id: Iebc5cfd5d0f43cb4435f81b641ccc91ee2d60fb5