wifi: ChannelScanDwellTime: fail on missing probe requests
The test expects to find probe requests in the pcap to ignore beacons
received before the probe request. However, in some cases the pcap does
not contain any probe requests and the test crashes because it directly
tries to access the first probe request.

The CL doesn't fix the underlying issue (why are there no probe
requests?) but tries to handle that error case cleanly to make it easier
to diagnose the issue.

BUG=b:206846478
TEST=./platform/tast/fast_build.sh -T

Change-Id: I37a006235cf1916a983e73c83fc633a9d97b0ecc