(On hold experimental) refactor: Enable Lacros in chrome.New with custom options
Currently lacros fixture is required to enable lacros for testing.
However, there are a number of the Tast tests that don’t use fixtures on
purpose for their own reasons and we may not want to add the fixtures
anyway and change test code more than necessary to migrate them to Lacros.

This CL will do:
- make it easy to enable Lacros without having to add
lacros fixture
- provide some abstraction of some chrome.Options commonly needed to setup
Lacros and avoid moving the flags around everywhere
- support various lacros availability: LacrosPrimary, LacrosOnly
- support various lacros source: Deployed, Rootfs, Omaha, External
- add example tests for a proof-of-concept

Note that it should be a no-op to existing ash-chrome tests as long as
LacrosEnabled is unset.

See http://go/lacros-tast-refactor-proposal for the details.

BUG=b:222242545
TEST=tast run ${DUT} example.Browser*
TEST=tast run -var=lacrosDeployedBinary=/usr/local/lacros-chrome ${DUT} example.Browser*
TEST=Run CQ regression checks
TEST=  tast run ${DUT} ("group:mainline" && "dep:lacros" && !informational)
TEST=  tast run ${DUT} ("group:mainline" && "dep:chrome" && !informational)
TEST=  tast run -var=lacrosDeployedBinary=/usr/local/lacros-chrome ${DUT} ("group:mainline" && "dep:lacros" && !informational)

Change-Id: Ie0f2aad5a74b6a7efc7c00c7391810a09d1fd5ff