Switch to using account pool
appsplatform.CheckSkuPurchase tests started to fail recently due to authentication error: "error": "rate_limit_exceeded"
This error occurs if there are multiple logins performed for a given user account within a short period of time.

Switch to using an existing gaiaPoolDefault account pool (see go/tast-ota-pooling for the details). With 20+ user accounts in the pool this should alleviate the issue.

BUG=b:219596236
TEST=tast run dut appsplatform.CheckSkuPurchase.vm

Change-Id: Iee3af49facf8b68ee75558ba12c8e1683c1b294a