Crostini: Test Gedit install and uninstall via terminal.
This test checks the termina container size before
and after gedit install, and after gedit uinstall. The test
also  uses Acuiti to interact with Gedit ui elements, and the
Gedit Icon in the Shelf.

This automates the following test case:
https://testtracker.googleplex.com/testplans/testcase/detail/2877157?id=18920&revision=199

Please note comment#3 in the bug on expected container size.

BUG=b:217091041
TEST=Ran on hatch (Akemi) DUT and remotely on lab Atlas boards.

Change-Id: Id9d4290463782790c261628afe797dd1b9959059