tast-tests: Remove necktie and unmanaged enterprise.ARCProvisioning
CloudDPC is the only device owner and DPC possible for ARC so testing
accounts with necktie enabled isn't useful as necktie cannot operate
without their own cloud dpc app.

Unmanaged is same as managed test since they are both using a managed
user and there is nothing else to interfere with the provisioning and
installing of apps.

Deleted the redundant variants.

BUG=b:186274474
TEST=tast run DUT enterprise.ARCProvisioning

Change-Id: I014ee19dc141e1470261081cc66370ca85e76a82