video: add code and cases for reading video playback "roughness"
Media DevTools has information of video playback roughness, i.e.
the distance between ideal playback and what's observed. This
CL adds code to track that, and specific use cases for that (note
that we have to use crosvideo test vectors bc the ones available
under perf/* are 10s, hence too short for roughness to be
calculated).

BUG=b:171913706
TEST=tast run -ephemeraldevserver=false <IP> video.PlaybackPerf.*_long
on volteer

Cq-Depend: chromium:3518745
Change-Id: Ief7e0da26174253e1ae7896b615b53bd65bdf5bd