[SmartLock] Add E2E tests for SL revamp settings CUJs
Add tests for enabling/disabling Smart Lock in Settings and checking
whether Smart Lock shows up on the lock/login screens.

This test is very flaky. When disabling Smart Lock for sign-in,
logging out/in twice does not make the setting take effect, but
doing the same thing manually works fine. I'm planning to investigate
this and fix in a followup.

BUG=b:223435238
TEST=tast run 'smartlock.Settings'

Change-Id: I62e84c15459e70ce4ed08bff7c81d866c9c50e1d