tast: add arc_p version of multivm.login test
Existing test multivm.login has no history of running on
ARC++ (container). Adding the new flavor, so that it buildst this
history going forward, and it gets ready to issue data for
compating ARC++ and ARCVM on upcoming models we release.
This will also help local testing.
Note that the necessary infrastructure for talking to resourced
landed in 14092.0.0 ( per
https://crosland.corp.google.com/cl?q=chromium:3020705 ).
It works okay on Milestone 92 as well - but using milestone 93 to be safe,
to keep the test from hanging on older builds.
Tested on milestone 90, graceful failure.
Chose that milestone as it marks the baseline for the ARCVM memory
improvement efforts.

BUG=b:217276698

TEST=tast -verbose run __dut__ multivm.Login.arc_p

Change-Id: I0e88de20e74fc54fc7959b91e67c5e23bec77b8c