tast: add arc_p version of multivm.login test
Existing test multivm.login has no history of running on
ARC++ (container). Adding the new flavor, so that it buildst this
history going forward, and it gets ready to issue data for
compating ARC++ and ARCVM on upcoming models we release.
Also created a short timeout for connecting to DBUS, to fail
faster if running against older system, or if there's a regression
in the connection to resourced.

BUG=b:217276698

TEST=tast -verbose run __dut__ multivm.Login.arc_p

Change-Id: I0e88de20e74fc54fc7959b91e67c5e23bec77b8c